update for new fcl (noetic) + float

+++ include/fgpg/calc_area.h
@@ -9,19 +9,19 @@
 #include <boost/geometry/geometries/polygon.hpp>
 
 
-double getGraspDistance(const Eigen::Isometry3d& transform, const FCLGripper& gripper, const std::vector<TrianglePlaneData>& planes)
+float getGraspDistance(const Eigen::Isometry3f& transform, const FCLGripper& gripper, const std::vector<TrianglePlaneData>& planes)
 {
   auto n = gripper.getPalmNormalVector(transform);
   auto o = gripper.getPalmOrigin(transform);
-  double min_distance = std::numeric_limits<double>::infinity();
+  float min_distance = std::numeric_limits<float>::infinity();
   int index = 0;
   for (auto it = planes.begin(); it != planes.end(); it++)
   {
-    Eigen::Vector3d p;
+    Eigen::Vector3f p;
     calcLinePlaneIntersection(*it, o, n, p);
     if (pointInTriangle(p, *it))
     {
-      double dist = ((o-it->points[0]).transpose() * it->normal);
+      float dist = ((o-it->points[0]).transpose() * it->normal);
       if (dist < 0.0) continue;
       if (dist < min_distance)
       {
@@ -42,25 +42,25 @@ double getGraspDistance(const Eigen::Isometry3d& transform, const FCLGripper& gr
 // {
 // public:
 //   GraspAreaCalculator();
-//   double calcArea(const Eigen::Isometry3d & transform)
+//   float calcArea(const Eigen::Isometry3f & transform)
 //   {
 //     auto f1 = gripper_.getFinger1PlanePoints(transform);
 //     auto f2 = gripper_.getFinger2PlanePoints(transform);
 //     auto p = gripper_.getPalmPlanePoints(transform);
 
-//     std::vector<Eigen::Matrix<double, 3, 5> > planes;
+//     std::vector<Eigen::Matrix<float, 3, 5> > planes;
 //     planes.push_back(f1);
 //     planes.push_back(f2);
 //     planes.push_back(p);
 
-//     typedef boost::geometry::model::polygon<boost::geometry::model::d2::point_xy<double> > polygon;
-//     typedef boost::geometry::model::d2::point_xy<double> point_type;
+//     typedef boost::geometry::model::polygon<boost::geometry::model::d2::point_xy<float> > polygon;
+//     typedef boost::geometry::model::d2::point_xy<float> point_type;
 
 //     for (auto & plane : planes)
 //     {
-//       Eigen::Vector3d normal = plane.col(4);
-//       Eigen::Vector3d v1 = plane.col(0) - plane.col(1);
-//       Eigen::Vector3d v2 = plane.col(0) - plane.col(2);
+//       Eigen::Vector3f normal = plane.col(4);
+//       Eigen::Vector3f v1 = plane.col(0) - plane.col(1);
+//       Eigen::Vector3f v2 = plane.col(0) - plane.col(2);
 
 //       polygon triangle_polygon, gripper_polygon;
 //       // triangle.outer().
