+++ include/fgpg/hsv2rgb.h
@@ -4,9 +4,9 @@
 
 /*
  * H(Hue): 0 - 360 degree (integer)
- * S(Saturation): 0 - 1.00 (double)
- * V(Value): 0 - 1.00 (double)
+ * S(Saturation): 0 - 1.00 (float)
+ * V(Value): 0 - 1.00 (float)
  * 
  * output[3]: Output, array size 3, int
  */
-void HSVtoRGB(int H, double S, double V, int output[3]);
+void HSVtoRGB(int H, float S, float V, int output[3]);
